2024, Tuesday 9th of July
Today I've done making heat box, box graph, and scatter graph. The data that I analyzed is the breast cancer data from Kaggle. I first tried to make a scatter graph and then a heat box to find the corresponding data. Then, from what I got, I used the highly corresponding data to create a box graph and look at the similarity. On the x-axis, the value that I used was consistent which was the diagnosis so that I can look up the relationship between the y-axis-data and the result.